First slate of changes to the site to support CARTO -> ArcGIS map version migration.
The main map is going to be migrated from CARTO to ArcGIS. This necessitates comprehensive changes to the site's documentation and visuals for consistency (despite much of the map's core functionality being the same). This commit marks the first of a slate of changes that are required for this migration. Following subsequent testing future changes will be made before the ArcGIS version will be merged with the "production" gh-pages branch.