Additional changes to support ArcGIS version migration.
Additional changes have been made (primarily to the tutorial and map pages) to reflect the differences in the core functionalities of the ArcGIS version of the CSOM map. This state of the site at this commit will be shared with the rest of the group for review and error checking.